Moved `NoManufacturerCluster`
From `zhaquirks.tuya.mcu` to `zhaquirks.tuya`
